add HealthWho
this commit also introduces household_questions. these are
questions in which we want to show all members of a
household as an option to select. when selected, a
boolean on their member record is updated.

both household and member questions overwrite
the standard 'update', so i've also created a special
step subclass for these: ManyMemberUpdateStep.