Refactor use of `User`
**WHY**: Before, we were creating a `User` record in the database with
bogus data for each new SNAP application.

This PR moves us toward what GCF does, which is saving the `id` of the
current SNAP application as a session variable and determining relations
based on that information.

The benefit of this is that we are standardizing on one approach across
apps and not creating meaningless database records.

This app also kills MOM.