Add and configure mailcatcher
In order for mailcatcher to work, we have to boot it up alongside both
the rails and worker processes. A small wrinkle attached to this is that
we don't want this process to start via the main, canonical, Procfile.
We'll need a separate, dev-specific, Profile that we can run locally -
`Profile.dev`. In order for this procfile to run in the dev environment
we can add a .foreman file that will point to the other file -
`Profile.dev`.

* Move mailcatcher to development gem group
* Configure mailcatcher in config/environments/development.rb