Add sfax, corresponding fax adapter and test
The sfax gem has been taken over by another project so pointing to the
CFA fork that the GCF team has set up is the best bet.

This commit adds an adapter class to interface with the SFax gem and
tests that it is being called correctly, using the credentials places in
the Rails secret yaml file.

NOTE: Adding faraday is necessary, for some reason to get things to
function reliably.