Add env-specific shubox id's
Shubox works by mapping the domain the uploads are originating from to
the corresponding bucket on S3. It's how it, shubox, knows where to
upload files TO. Each domain also has its own custom JS file that must
be loaded on the website.

The one exception being domains that are analogous to,
and including, localhost - 127.0.0.1, 0.0.0.0, etc. Those will all "just
work" if you use the sandbox javascript file. No S3 configuration is
needed as it will use a bucket Shubox has set up already.

This commit pulls each environment/domain JS files' ID's into the
secrets yaml file so that, depending on the environment, the browser
will request the correct JS file to use for, and with, the domain.