Switch from HAML to ERB entirely
Moving foward with ERB allows anyone joining the project to not shoulder
the extra work to know the ins and outs of using HAML. While it's not a
tremendous effort it's still an extra amount of work. This also sticks
with the convention and toolset that the GCF team uses in their
codebase.

* Thanks to @danchoi's [Herbalizer project], this was a relatively trivial task
* Change the `icon` helper method to use `content_tag` instead of `haml_tag`
* Remove HAML and HAML linter from Gemfile

[Herbalizer project]: https://github.com/danchoi/herbalizer