Update docs upload UI to be more like GCF
Using a UI that's more analogous to GetCalFresh gets us a little further
along in the development without re-inventing another solution to a
design they've already tackled.

This commit re-uses some of their (s)css and extracts some of the views
out to partials that can be used with a templating language like
handlebars, which is also included here.

Using handlebars allows us to render some html we already use for the
documents collection - therefore not needing to escape into a large JS
string.