Add bundler-audit to default rake task(s)
The bundler-audit gem allows us to check that all currently installed
gems in the bundle are not versions that have had a security issue
reported as a CVE.

This commit adds the rake task `bundler:audit` to the task pipeline in
addition to rspec and rubocop. The test suite will *not* pass if a gem
has an open CVE in the bundle.