Refactor flash-partial code
The existing display/dismiss flash partial code is tricky within the context of
the new michigan benefits app so I've made some changes to allow things to fit
in, for the time being, until there's a more robust visual treatment.

This commit changes things to loop through all flash keys and values and
display the corresponding html. It keeps the ability to dismiss the flash
notices albeit on a slightly superficial level (read: using the `:target`
pseudo-class in css).

![flashes](https://www.evernote.com/l/AUpXEy8RGFFMYZHt0KvEyN5hVvdSyL1vj8sB/image.png)