Fix Fax model to ensure successful fax sending
After testing on the console with some sample operations against the
live SFax service there were clear issues with it going through
successfully. After some source diving and inspecting the HTTP requests
it looks like a third parameter is required (an undocumented
requirement) for it to work.

After adding the "recipient name" to the method and executing the
command, it worked.