Redirect home if no application is in session
If someone tries to go to one of the steps (steps that are NOT the very
first step) then they should be redirected to the homepage. Once they
start the application process and create a new application, they will be
able to go back and forth along all of the steps.