Use headless chrome instead of poltergeist (#123)
When our JS-enabled tests run we should use (headless) chrome and
chromedriver as the driver. This removes the dependency of poltergeist
but adds the need for chromedriver. As a result:

1. Add the installation of chromedriver to our setup script
2. Add the installation of chromedriver to circle.yml

Because the platforms we test on install chrome(driver) in different
locations this commit adds a `FindChrome` class that will locate chrome
and chromedriver in either the most common locations, or a location we
can specify with the ENV variables `CHROME_PATH` or `CHROMEDRIVER_PATH`.
If they're not found - raise an exception.

Default JS driver is `:headless_chrome` but can be changed to `:chrome`
in `rails_helper.rb` if we want to watch those tests in a real chrome
browser.

Also:

* Update feature_helper to work with this driver. For some reason
capybara was complaining about the previous api usage - something about
how that wasn't implemented. ¯\_(ツ)_/¯