Add sample applications with built in db:seed task
For things like testing the web driving, it is required that we have a
"finished" application that can be used for that work.

This commit adds sample data from one end-to-end submission with the
bare minimum. With regards to the data - nils, empty arrays, or anything
that was "barren" was removed.

This will also add a more fully fleshed out snap application.