Click through benefits and fill in personal info
There are two benefit overview pages that the driver can click through
before coming to a more involved form page where they enter personal
information:

* Name
* Gender
* Birth date
* Address/Location
* Phone number

If the client has an address that MIBridges considers an invalid
physical address it will re-display that personal information page and
confirm that the client *still* wants to submit with that street
address. If so, this will confirm that this is, indeed, the address we
would like to continue with.