Fix the document uploading UX
When adding documents to an application, there is a duplication of
elements - the first being the "preview" and the second being the final
uploaded image/file.

When an upload is sucessful or results in an error, the preview element
is removed from the DOM.

As files are being uploaded, the "done uploading" button is able to be
clicked before things are completely finished uploading. This will
occasionally result in the form being submitted before a/some file(s)
are completely uploaded.

The "DONE" button is disabled until uploading is finished.

![](http://g.recordit.co/Yqu8Kw9eou.gif)