Refactor js that warns when exiting application
Since we have several new links that need to be accounted for -- namely
privacy and terms -- this JS can/should be changed to be more amenable
to changes.

This update creates a whitelist of strings that the js will look for
and, if the link contains any of those strings, it will not consider it
an external link.

TLDR, if these are substrings of something you click, the warning
message won't pop up.