Do not display newly uploaded file post-upload
As a step towards locking down access to files uploaded to S3 we must no
longer display the successful image URL after the file is uploaded.
Instead, let's just keep the preview element there and let the JS show
the contents of the file that way.

Also pull in the work Rachel and Zee did to display upload errors inline
within the preview element. Including:

* Allow files up to 8Mb in size
* Don't show the warning by default