Add common "Click Next Page" super class
There are more than a fair share of "review" or "summary"
pages on MI Bridges that need only to have a "Next" button
clicked. To represent the behavioral needs of these pages
a super class in the form of a "ClickNextPage" class is
made in this commit for other classes to inherit from.

This allows us to

1. Change the behavior in only one place if need be
2. Give the inheriting classes their own descriptive class name so viewing the queue of pages in `driver.rb` self-documents.