Abstract the conditional checking of boxes
The questions on these pages are constructed in a slightly confusing manner:

The best way to scope each question is to hook into an aria attribute
and check the boxes by executing javacript against a selector that works
thanks to jQuery. That is how `check_in_section` works. For a better
example -- and please note that `aria-labelledby-section-name` is the
value for the html property `aria-labelledby`:

```
check_in_section("aria-labelledby-section-name") # checks "No one"
check_in_section("aria-labelledby-section-name", if: predicate?, name: "John") # checks the box next to "John" if the predicate method returns true
```