Remove default checkboxes from PDF
**WHY**: these checkboxes were checked by default before since, if there
were a value, we'd always know that it was for a specific interval. But,
this default behavior was confusing MDHHS workers because it would show
up even when the value was nil. Eg: no court ordered expenses, but still
a checkbox next to "monthly" in the court ordered expenses category.