Downloading public *and* private S3 files
Soon all files being uploaded to S3 will be private. Requesting them
directly via their S3 url will usually (unless it's in development env)
not work.

As a result, we need a fallback in those cases when those files are
private. That fallback is the AWS SDK. Using the requisite credentials
we can download them that way and use them in our PDF creation workflow.

This commit extracts some of the work done previously in
`VerificationDocument` and houses it in `RemoteDocument`.

Also:

* Add dotenv gem for cases where local servers are run without Foreman
* Add FileMagic gem to inspect file mimetypes
* Install libmagic in `bin/setup` as it is needed for FileMagic