Tighten up that release diff (#234)
Our deploy script opens up a giant diff every time because the production server hadn't had a real git push in a long long time. This wasn't very usable because it had lots of noise and stuff outside of the actual changes between production and staging. This tightens it up to comparing the latest production deploy to the latest staging deploy.
* https://trello.com/c/fZg67unC/306-improve-readability-of-deploy-diffs