Handle failed attempts getting non-existent S3 object
In the cases where an object in S3 no longer exists we need to handle
the effects that has in our system. This means we should NOT try to
attach this non-existent document.

This commit rescues the error from the AWS SDK and bubbles up the
resulting nils to the DHS1171Pdf class where it does not include them if
they are nil.