Pages for mailing addr, pregnancy, self-employment
Page objects for "Mailing Address", "Pregnancy Information", and
"More about <PERSON>'s Self Employment". A bigger change as a result of
the "Self-Employment" section is that the string we're testing against
in the H1 can now be dynamic - it can reference a member of the
application as per this example.

As a result this commit will check for both equality of the string, and
whether there is a match in the chance we make `TITLE` a regular
expression.