Add "More about <user>'s income'" page
In addition, add more information about when we hit the specific error
when a page's title is not found in our registry of page objects.

Why? Because it will tell us exactly what the page title is that is
being searched for, therefore allowing us to add a page that adheres to
the pattern.