Adjust payment intervals to match MIBridges
Now that we're filling out the page in MI Bridges that asks for this
information we know where/what we need to select and put in. This commit
makes the changes, pushes these options into an array that can be used
for both validation AND the values we pass to the form builder in the
view.