Add default response for Child support question
**WHY**: I didn't catch this before because MI Bridges only shows this
question if there is someone under 18 in the household.

The tricky thing about this is that we know if *anyone* receives child
support income in a household, but MI Bridges wants to know which child.

After you select a name, it follows up with a queston about the amount
of child support income for each name selected.i

Options are:
- always say "no one" (easiest)
- always select first child (second easiest)
- collect more granular child support data in MI Benefits so we can fill
this out accurately (hardest)

In order to know which is best, it would be good to get context on the
importance of this question for the county workers.