Add cash on hand page object
Cash on hand page shows up amongst all the income questions so this
takes care of that.

Also changing the regex in the title to capture the user's name allows
us to not have to redefine that regex below in the "More about Job
Income" page, just like in this "cash on hand" page.