Real estate and vehicle information pages.
The real estate and vehicle information pages are tricky in that we
don't ask for more information about *what* type of vehicle or
house/home/domicile the client owns. We just know they have one or do
not. As a result we'll be as generic as we can and continue on with the
driving process.

Also move "JobIncomeMoreAboutPage" to "MoreAboutJobIncomePage" so that
it adheres to our "MorAbout*" naming convention.