Allow skipping instance limitation for some pages
[According to Jessie] there are instances (pun intended) where some page
objects are created more than the number of times we're limiting. As a
result we should allow these page objects to opt out of the instance
check.

This commit adds that ability by defining a class method for the page
objects we expect could run into these situations - `.skip_instance_limitation`.

[According to Jessie]: https://github.com/codeforamerica/michigan-benefits/pull/287#discussion_r141438656

Also - refactor related initializer contents to be a little easier to read.