Drive medical bills/expenses pages
The medical bills page provides a collection of check boxes to state
what sort of medical bills the client pays. Subsequent pages are shown
depending on what was checked on that medical bills page.

This commit handles those pages. It also adds ALL possible medical
expenses that correspond to the MI Bridges options into the
"complete_application" object we create in the db seed task.

Note that all of the derivitive "more information" pages are "click
next" pages since we do not collect any more information beyond whether
or not the client is paying something for them.