Save error info for all failed driving attempts
When driving errors out, we should get some visibility into _why_ things
failed. The information we have available upon exceptions being caught
are the exceptions themselves, the message property in the exception,
the current page object we are attempting to run, and the contents of
the page we are on currently.