Display driving-related errors in the dashboard
After an application has been driven and runs into a problem, we save
that information in the `driver_errors` table. To get better vsibility
into these instances we should be able to see what's happened in the
admin/dashboard.

Snap applications now have a different color background based on the
status of their previously attempted drives.