Record when apps (and its errors) are driven
When we create and drive an application we are saving the information
about the drive to a `driver_applications` record. When these drives
occur we should be recording the date and time WHEN the drive is
occurring. In addition, the associated errors should have the same time
set to them so that errors can be grouped by the time they were all run
by its parent `DriverApplication` object.