Add chrome and chromedriver buildpacks to app.json
When deploying to a review application we should have the chrome and
chromedriver buildpacks available in order to make sure the web driver
code will run correctly.

Also:

Use ENV from by chrome buildpack to find binary

In the off chance the buildpack changes the path it's best to use the
ENV that is provided by the buildpack directly. A hardcoded path is
liable to cause breakage if the buildpack changes where the binary
lives.