Display exports for snap apps in admin/dashboard
In order to get more visibility as to *where* and *when* applications
are sent to the offices in MI we can display the exports in the admin
in the context of their parent snap application.

This commit adds the resource to the admin and adjusts the seed task to
create some fake data for local testing and demo'ing.