Add good and strong migrations gems
There are two reasons why I'm suggesting we add these gems.

1. We're running into odd issues with the schema file being generated
and causing diff issues when something like overcommit is run. This has
prevented several people from rebasing a branch and can be mitigated if
the schema has a more predictable diff, and is generated only when it's
necessary. `strong_migrations` helps us with that. See:

  * https://github.com/ankane/strong_migrations#schema-sanity
  * https://github.com/ankane/strong_migrations#faster-migrations

2. We're continuing to generate and run many new migrations - therefore
the constraints inherant with these gems will help us move more
carefully, and in some places with more speed. Constraints are good.

In this commit I've added initialization configuration as suggested by
the README in the strong_migrations project, and have generated a brand
new schema dump.

See here for more information:

https://github.com/ankane/strong_migrations