Add medicaid_application model & back it w/form
Medicaid information must be stored, of course, so this commit
introduces that first migration and model to get things started.

In addition, an update to the form builder here allows us to abstract
the common pattern of the "yes" and "no" buttons on the Medicaid side of
the application. Clicking one or the other will pass the boolean value
to the update action.