Add yes/no form partials & assoc fields to steps
The views that ask for a yes or no answer can all display those forms
through one shared partial. These views have all been updated to pull in
that (new) partial.

The fields for each of those steps have also all been added to the
related step classes.

Last but not least, after looking at things, it probably makes more
sense to surface any HTML to the views, that CAN be displayed in the
view. The form builder method is just an obfuscation at this point if we
can handle this in the partial.