Add test coverage for controllers re: skip logic
This commit adds some tests to back the recently created medicaid flow
controllers. The logic around skipping to the next page needs some
robust test coverage to ensure the logic does not break in the future.

Also move the `next_path` method to the public method space so we can
call it from the specs to test that they are skipping to the correct
"next" path.