Only export apps when they're sufficiently done
If there is no signature on a snap application then they should *not* be
sent to the offices in MI. This can occasionally happen if someone is
skipping steps or bouncing around the application and not filling
everything out sufficiently.

This commit introduces a guard clause in the final success controller
that will bail out of the Export/Driving work if the application is not
"done" enough. For now this checks for only the signature being present.
In the future it stands to reason that there will be further conditions
that define whether or not an application is "done".