Allow multiple members for Medicaid application
For single member flow, it made sense to save some user details on the application.
Now with multiple members, it made sense to move that information from the application and into the member class. To do this, we made applications (Snap and Medicaid) a polymorphic association on the Member class.