Log info about email to both export & logger
When an export is executed, the block we pass to it can accept a logger
object which, when written/outputted to, will save to the export object
AND log to stdout.

Our email job has not been doing this so this commit changes it to do so
with not only the receiving office, but the id of the client who is
submitting it.