Silence deprecation errors re: AR + attr_encrypted
When running the test suite we see a LOT of instances of the following:

```
DEPRECATION WARNING: user_id is not an attribute known to Active Record.
This behavior is deprecated and will be removed in the next version of
Rails. If you'd like user_id to be managed by Active Record, add
`attribute :user_id to your class. (called from block (3 levels) in <top
(required)> at ./spec/models/snap_application_spec.rb:235)
```

This is due to the `attribute` class method not liking being given an
array of symbols as a parameter. Passing each attribute to the
`attribute` method fixes this.