Better prompt when asking abt household pregnancy
Because the pregnancy question is specific to how many females are among
the members, the question for this should be catered to several
different contexts, or states:

1. Multiple members. One female. That female is you. - "Are YOU pregnant"
2. Multiple members. One female. That female is someone else - "Is SHE pregnant"
3. Multiple members. Several females. "Is ANYONE pregnant"

When there is a single female in the household that's pregnant we should
not show them the page where we ask who. It's a given that if there's
only one female in the house, and someone is pregnant - then she is that
person.

This commit revises the view and the locale to specify how those
questions start and personalizes them accordingly.

A view spec is added as well so we don't exercise the entirety of the
application in a feature spec.