Only show application-specific footer in application flow
An error was occurring on static pages because we were trying to access the current application, which isn't available for static pages. This makes that footer only visible on pages where we expect an application to be present.

TODO: Currently we don't show the application type on the first step, because an application isn't made until the first step is submitted. Does not cause any issues, but may want to address.

Signed-off-by: Christa Hartsock <christa@codeforamerica.org>