Collect insurance type per client, per page
We want to loop through each user and ask them what their current
insurance plan is through. To do so we must break out of our current,
codified, patterns. We have to start with *A* member, and move to the next
member after submitting and saving, until we've reached the last user.

To complicate things, we have to be able to go *back* from the next
page/section and allow the client to re-enter adjust the insurance type
if they decide to go back.

To accomplish this we need to create the concept of a _current_member_
and determine the next and previous members and the accompanying paths.