Make documents and paperwork more generic
That is, instead of referring to it as one or the other, make it a
generic concept and call them "uploadables" as a basic abstraction.
Allow the views and the two benefits applications to call them something
different -- whether that's documents or paperwork.

QUESTION: This is, for all intents and purposes, a user-facing commit.
Only changing what we are showing the clients. On the database level
would it make sense to then change the field from "documents" or
"paperwork" to "uploadables" as well.