Use `display_name` instead of `full_name` to show member names
**WHY**: in the future, we may want to change how we do this - maybe we
show first name and DOB, for example, or first name and last initial.
This way of naming the method makes it more flexible.

I also took this opp to move to use of `display_name` in places where we
were just concatenating first name and last name.

I also also took this opp to titleize first name and last name. :)