Extract step attrib reconciliation to own object
... and make the form builder behave accordingly when an attribute is of
type "array".

The majority of the attributes on our steps are mapped plain vanilla
literals - strings, integers, bools, etc. On the VERY rare occasion
there are fields that map to arrays of those literals. For example -
documents and paperwork.

Documents and paperwork share one partial that has explicitly
overwritten the value of the "name" option in the form input field.

For example. A normal form field comes out like so:

`<input type="text" name="step[first_name]">`

But when we would like N number of values for a field that maps to a
*array* column in the database, we'd like the html instead to be as
follows:

`<input type="text" name="step[monthly_income][]">`

Before this commit, the only way to do that would be to override the
name option with something akin to:

```
<%= f.mb_money_field :monthly_income,
  "Job N",
  name: "step[monthly_income][]" %>
```

After this commit, the form builder will be smart enough to know that
that attribute is hash key for an array/collection.

***

Also!

In addition to the above, the work that's been done inside the base step
class to munge/extract information about the attribute definitions that
are provided by the `step_attributes` class method has been extracted to
its own class to do the work of extracting the string values that are
passed to `attr_accessor`.

Lastly - a method to ask "is this attribute a symbol or hash key" is in
this attributes class to allow the form builder to decide whether the
form field needs to allow for multiple values (for arrays) or not (for
object literals)