Form builder always treats fields as array of vals
Instead of asking whether a form field is a key and value *OR* a key and
multiple values, this commit will change the value to be an array,
always, and loop through it when creating the label and input.

Using Array.wrap a single value (eg: "String") will be changed to
["String"] and can be looped through the same as the less common
instance of when a column is an array of literals.

It's confusing. Basically normalizing the "value" to *always* being an
array (sometimes one, sometimes many) makes it easier to create the form
fields.