Allow prepending HTML (similar to append_html)
There's a named parameter called `append_html` that allows providing
some html that can be included just after the input tag. Why? Because
we want to inject something still inside the `fieldset` tag that we wrap
around the label, input, etc.

This commit also allows us to interpolate the index of the value/input
as its being rendered. The use case for this is for when we are
outputting N number of inputs that correlate to values in an array type
value. Sometimes we want to say this is the 1st, 2nd, 3rd item in the
label, or in the prepended/appended html text.