BUGFIX: Show addresses correctly in 1426 PDF
* Explanation:

The PDF for the Medicaid application (1426) displays a home address and
a mailing address if that address is different.

Previously, we were not rendering these properly because our application
collects residential address and then mailing only IF they are
different, but our ruby logic returned mailing address and residential
address only IF they are different. This makes sense for how the SNAP
flow works, but resulted in not returning correct data for Medicaid.

To clarify the behavior in tests, we removed the `address` factory so
that only `residential_address` or `mailing_address` factories can be
used.

* [Finishes #152564632]

Signed-off-by: Jessie Young <jessie@cylinder.work>