Ask for tax filing status for tax household members
Updates primary member's tax relationship (filing status, effectively) to "Single"
if noone in household is filing jointly, and "Joint" if one other member of household
has indicated they are filing jointly.

Also introduces pattern of redefining step members in the controller, so that the controller
is scoped to only consider applicable members (in this case, tax household members).

https://www.pivotaltracker.com/story/show/152561623
[Finishes #152561623]

Signed-off-by: Paras Sanghavi <paras@codeforamerica.org>