Fill in "Other" as pay interval when "Hourly"
* Hourly is not an option in MiBridges, but it is in our SNAP flow
* Pay interval is a required question for employed members in MiBridges,
so whenever they've reported it to us as Hourly, the flow fails on this
page
* This issue presents a larger product question, but for now the way we
can move past it and continue to drive apps where there is an employed
household member is to select "Other" in this case.
* Tagging the product / design story in here so we can un-do this
default select when we have more clarity on the longer term fix.
* [#153378349]
* I confirmed that this works with the SNAP id referenced in the story