Reset exception handling within page loop
When we moved exception handling a level higher, we lost the ability for a drive to pick up where the exception was raised. This resets it to the original behavior and adds backtrace logging. It does not handle properly catching and logging problems with logging errors when we hit TooManyAttempts error, which moving exception handling up was meant to solve.

[#152982028]

Signed-off-by: Ben Golder <bgolder@codeforamerica.org>