Eager load lib directory
In order to make lib available to the Rails application in all environments, we need to either manually require certain files or eager load them (production doesn't autoload).

While tracking down the chrome and chromedriver loading, we noticed that there were two places where we were trying to load the driving code previously--one in application.rb, which wasn't working correctly because it was autoloading (ignored in prod), and another in initializers/extensions.rb, which was loading the app correctly. Decided to instead do it in one place.

More details on this here: https://stackoverflow.com/a/19852844

[#152649354]