Run accessibility checks in feature specs using `axe-matchers`
* https://github.com/dequelabs/axe-matchers
* We are doing an accessibility check for each page using the
`proceed_with` method, which is really just an alias for `click_on` plus
the accessibility check.
* In some cases, doing the accessibility check scrolls the page in such
a way that the button is underneath the header element on the page,
which Capybara does not like. So, we need to scroll to the top on those
cases to bring the button to the foreground of the page.
* [Finishes #152982764]
https://www.pivotaltracker.com/story/show/152982764

Signed-off-by: Luigi Ray-Montanez <luigi@codeforamerica.org>