Update legal agreement pages for a11y
* Fix invalid HTML as detected by WAVE
* Fix contrast issue
* Contrast issue brought up issue with our overrides, which was that we
load our overrides last in the manifest file so that they take
precedent, which means that our variable overrides were not loaded until
after all of the base styles. Now, we are loading our variable overrides
above the base styles so that the base styles that use variables are
using our overrides rather than the original colors.
* [Finishes #153764205]
* [Finishes #153764168]