Fixes "MiBridges: Closing the driver causes an error" (#617)
* The final step was calling `close`, so when the teardown method in
`Driver.rb` tried to close the browser, it got an error because the
driver had already been closed
* This also fixed another bug that popped up recently where the Privacy
PIN  page option was not being selected properly. The error showed up as
`Selenium::WebDriver $ is not defined`. Fixed by calling `find` on the
element so that it is loaded (Capybara waits for it) before clicking on
the id.
* [Finishes #153571372]