Convert max attempts for delayed jobs to integer
When a delayed job failed, a problem with our configuration (storing the number of max attempts as a string) was causing the worker to crash when it tried to reschedule. As a result, the job never got its number of attempts updated and unlocked, so a failed job would continue on forever, blocking any new delayed jobs that got queued. See this ticket for a symptom: https://www.pivotaltracker.com/story/show/153996535.

This PR does two things:
* Converts our configuration to an integer, which is what delayed job expects. This prevents the worker from crashing when it tries to rescheudle a failed job.
* Prevents delayed job from deleting failed jobs (the default). This will allow us to see what jobs are regularly failing in the future—otherwise, they just disappear into the ether. Here's a good article recommending this as a best practice https://www.sitepoint.com/delayed-jobs-best-practices/

[Fixes #153996535]