Create script and environment for debugging with production data
Sometimes we need to troubleshoot problems using production data. This script and related Rails environment allows us to connect to the remote production database using our local code to ease troubleshooting and keep data in the DB.

With great power comes great responsibility (and I'm open to not merging this into code base, but wanted to offer it as a point of discussion!).

[#154448184]