Redirect user after `/success` if they try to access page without app
Previously, if a user submitted a SNAP application and clicked 'back' after hitting the homepage, they would receive an error if they clicked on 'skip this step'. This commit addresses that by redirecting the user if there isn't a current application when they click 'skip this step', and adds a helpful flash message to let them know they've been redirected to the start of the application.

[Finishes #154097480]