Consolidate logic for detecting whether our app should really send applications
Right now we have two ways to determine whether an application should be sent to a real office--by the "APP_RELEASE_STAGE" environment variable, which routes applications to fake offices if it's set to "staging", and Rails.env.staging?, which is used to place a banner at the top of the site alerting applicants that it's fake.

To allow easier configuration (and potentially allow us to eliminate our staging environment), this commit consolidates the checking to the GateKeeper object that hooks into the existing APP_RELEASE_STAGE environment variable.

[#153856313]